Attempt with Bijoy
Today, our SPL course teacher asked us to create a program which can print all the even values from 1 to 10 without even using loops. She mentioned to use if-else statement. I tried to do it using array but because of some technical methods of if-else statement, I failed to do it in the class. After the class, I called my friend Bijoy and he gave me a brilliant idea to solve the issue and his idea really worked out and made a successful attempt. Hence, all the credit goes to him. We made it happen together on 8th September,2024. Let's keep this achievement.